split PartialValue code into its own file
